requirements.txt
Django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.7
mypy==1.2.0
django-stubs==4.2.0
